requirements.txt
spotipy==2.23.0
librosa==0.10.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
networkx==3.1
python-dotenv==1.0.0
requests==2.31.0
genius-lyrics==1.3 